PR comments: - rename functionalization.py - add detailed comment about caching
Also:
- rename TODOs
- format
