+++ tests/compile/backend.py
@@ -1,7 +1,8 @@
 from copy import deepcopy
-from typing import Callable
 
-import torch
+from torch import fx
+
+from vllm.compilation.inductor_pass import InductorPassType
 
 
 class TestBackend:
@@ -11,20 +12,20 @@ class TestBackend:
     It also saves the graph before and after the custom passes for inspection.
     """
 
-    def __init__(self, *passes: Callable[[torch.fx.Graph], None]):
-        self.custom_passes = passes
+    def __init__(self, *passes: InductorPassType):
+        self.custom_passes = list(passes)
         from torch._inductor import config
         self.current_config = config.shallow_copy_dict()
         self.current_config['force_disable_caches'] = True
         self.current_config['post_grad_custom_post_pass'] = self.post_pass
 
-    def __call__(self, graph: torch.fx.GraphModule, example_inputs):
+    def __call__(self, graph: fx.GraphModule, example_inputs):
         from torch._inductor.compile_fx import compile_fx
         return compile_fx(graph,
                           example_inputs,
                           config_patches=self.current_config)
 
-    def post_pass(self, graph: torch.fx.Graph):
+    def post_pass(self, graph: fx.Graph):
         self.graph_pre_pass = deepcopy(graph)
         for pass_ in self.custom_passes:
             pass_(graph)
