+++ tests/compile/test_functionalization.py
@@ -3,12 +3,12 @@ import torch
 
 import vllm.envs as envs
 from vllm import LLM, SamplingParams
-from vllm.compilation.functionalization import FixFunctionalizationPass
-from vllm.config import CompilationConfig
+from vllm.compilation.fix_functionalization import FixFunctionalizationPass
 from vllm.compilation.fusion import (FusionPass, find_auto_fn,
                                      find_auto_fn_maybe)
 from vllm.compilation.inductor_pass import is_func
 from vllm.compilation.reshapes import RedundantReshapesPass
+from vllm.config import CompilationConfig
 
 from .backend import TestBackend
